Readd CSS
Previously, the small button that appears next to password fields was
missing. The bold letter P from that button would occasionally be seen
at the bottom of pages with password fields. This was caused by the
removal of the CSS file in a cleanup in 003b4c9.

This commit readds this CSS file and its associated entry in the
manifest.